feat #92 : 폴더구조 변경 및 상단 가게정보 컴포넌트 분리완료
