+++ .env
@@ -1,1 +1,1 @@
-NEXT_PUBLIC_BASE_URL=https://bootcamp-api.codeit.kr/api/6-10/the-julge
+NEXT_PUBLIC_BASE_URL=https://bootcamp-api.codeit.kr/api/0-1/the-julge
