+++ src/components/detailPage/Card.module.scss
@@ -70,7 +70,7 @@
   border-radius: 20px;
   padding: 12px;
   gap: 6px;
-  font-size: 14px;
+  font-size: 12px;
   font-weight: 700;
   line-height: 17px;
   color: #ffffff;
