+++ src/components/detailPage/StoreNotice.module.scss
@@ -69,7 +69,7 @@
 .shop_contents span {
   background-color: #ffd43b;
   border-radius: 20px;
-  font-size: 14px;
+  font-size: 12px;
   font-weight: 700;
   line-height: 17px;
   color: #ffffff;
@@ -144,27 +144,4 @@
   font-size: 16px;
   font-weight: 400;
   line-height: 26px;
-}
-
-// 여기서부턴 "최근에 본 공고" 스타일
-.notice_container {
-  width: 968px;
-  margin: 0 auto;
-  padding: 60px 0;
-}
-
-.notice_container > h1 {
-  font-size: 28px;
-  font-weight: 700;
-  padding-bottom: 32px;
-}
-
-.card_container {
-  display: grid;
-  grid-template-columns: repeat(3, 312px);
-  gap: 32px 14px;
-
-  @media (max-width: 1028px) {
-    grid-template-columns: repeat(2, 312px);
-  }
 }
